move config file in config dir
