+++ src/discord/events/presenceUpdate.ts
@@ -1,6 +1,6 @@
 import { ActivityType, EmbedBuilder, Presence } from 'discord.js';
 import { BotsProps } from 'src/interfaces';
-import { LIVE_ROLE, STREAM_ALERTS } from '../../configs';
+import { LIVE_ROLE, STREAM_ALERTS } from '../../config/configs';
 import { LogEventType } from '../../enums';
 import { logEvent } from '../../utils';
